Rearrange and comment all utils functions
